requirements.txt
kivy==2.3.1
requests==2.31.0
kivy-garden.mapview==1.0.6